OSIS-6245 : Create a new test file
